Move some debug code blocks into their own files
